rename apiHelper to api to keep it simple
