+++ src/features/users/api/users.api.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from "axios";
-import { api } from "api/apiHelper";
+import { api } from "api/api";
 import { UsersEndpointsEnum } from "../constants/users.endpoints";
 
 const list = (): Promise<AxiosResponse> => {
